app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
boto3==1.35.0
sqlalchemy==2.0.35
psycopg==3.2.1
aiokafka==0.11.0
redis==5.1.1
python-dotenv==1.0.1
